requirements.txt
streamlit==1.45.1
tensorflow==2.12.0
pillow
numpy
pandas
